requirements.txt
Berlyne
django
celery
django-celery
kombu
